Converted App, Console, Dataset, Datasets containers to components, containers only have mapStateToProps and connect. Datasets is now DatasetsList
